Title: 시력 검사, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240320/시력 검사/eye-test.c
@@ -2,7 +2,7 @@
 
 int main() {
     // 여기에 코드를 작성해주세요.
-    int a,b;
+    float a,b;
     scanf("%d %d", &a, &b);
 
     if (a>=1.0 && b>=1.0){
